+++ var/httpd/htdocs/js/Core.Form.Validate.js
@@ -270,14 +270,22 @@ Core.Form.Validate = (function (TargetNS) {
         DateObject,
         RegExYear,
         RegExMonth,
+        RegExHour,
+        RegExMinute,
         YearElement = '',
         MonthElement = '',
+        HourElement = '',
+        MinuteElement = '',
         DateYearClassPrefix = 'Validate_DateYear_',
         DateMonthClassPrefix = 'Validate_DateMonth_',
+        DateHourClassPrefix = 'Validate_DateHour_',
+        DateMinuteClassPrefix = 'Validate_DateMinute_',
         DateCheck;
 
         RegExYear = new RegExp(DateYearClassPrefix);
         RegExMonth = new RegExp(DateMonthClassPrefix);
+        RegExHour = new RegExp(DateHourClassPrefix);
+        RegExMinute = new RegExp(DateMinuteClassPrefix);
         $.each(Classes.split(' '), function (Index, Value) {
             if (RegExYear.test(Value)) {
                 YearElement = Value.replace(DateYearClassPrefix, '');
@@ -285,16 +293,34 @@ Core.Form.Validate = (function (TargetNS) {
             if (RegExMonth.test(Value)) {
                 MonthElement = Value.replace(DateMonthClassPrefix, '');
             }
+            if (RegExHour.test(Value)) {
+                HourElement = Value.replace(DateHourClassPrefix, '');
+            }
+            if (RegExMinute.test(Value)) {
+                MinuteElement = Value.replace(DateMinuteClassPrefix, '');
+            }
         });
         if (YearElement.length && MonthElement.length && $('#' + YearElement).length && $('#' + MonthElement).length) {
             DateObject = new Date($('#' + YearElement).val(), $('#' + MonthElement).val() - 1, Value);
             if (DateObject.getFullYear() === parseInt($('#' + YearElement).val(), 10) &&
                 DateObject.getMonth() + 1 === parseInt($('#' + MonthElement).val(), 10) &&
                 DateObject.getDate() === parseInt(Value, 10)) {
-                if (Options.DateInFuture) {
-                    DateCheck = new Date();
+
+                DateCheck = new Date();
+                if ( MinuteElement.length && HourElement.length ) {
+                    DateObject.setHours($('#' + HourElement).val(), $('#' + MinuteElement).val(), 0, 0);
+                }
+                else {
                     DateCheck.setHours(0, 0, 0, 0);
+                }
+
+                if (Options.DateInFuture) {
                     if (DateObject >= DateCheck) {
+                        return true;
+                    }
+                }
+                else if (Options.DateNotInFuture) {
+                    if (DateObject <= DateCheck) {
                         return true;
                     }
                 }
@@ -312,6 +338,10 @@ Core.Form.Validate = (function (TargetNS) {
 
     $.validator.addMethod("Validate_DateInFuture", function (Value, Element) {
         return DateValidator(Value, Element, { DateInFuture: true });
+    }, "");
+
+    $.validator.addMethod("Validate_DateNotInFuture", function (Value, Element) {
+        return DateValidator(Value, Element, { DateNotInFuture: true });
     }, "");
 
     $.validator.addMethod("Validate_DateHour", function (Value, Element) {